automate_end_2_end/nice_scripts/requirements.txt
atlassian-python-api==3.25.0
certifi==2022.6.15
charset-normalizer==2.1.0
Deprecated==1.2.13
fuzzywuzzy==0.18.0
idna==3.3
jenkinsapi==0.3.11
lxml==4.9.1
multi-key-dict==2.0.3
numpy==1.23.1
oauthlib==3.2.0
pandas==1.4.3
pbr==5.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-gitlab==3.6.0
python-jenkins==1.7.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
six==1.16.0
urllib3==1.26.10
wrapt==1.14.1
